+++ src/App.tsx
@@ -7,7 +7,7 @@ import Button from 'semantic-ui-react/dist/commonjs/elements/Button';
 import Icon from 'semantic-ui-react/dist/commonjs/elements/Icon';
 import Dropdown, { DropdownProps } from 'semantic-ui-react/dist/commonjs/modules/Dropdown';
 import Sidebar from 'semantic-ui-react/dist/commonjs/modules/Sidebar';
-import { AddGame, Effectiveness, Export, Footer } from 'components';
+import { AddGame, Effectiveness, Export } from 'components';
 import { BadgeEditor } from 'components/Badges/elements';
 import AppRouter from 'routes/AppRouter';
 import useStore from 'store';
@@ -24,6 +24,18 @@ function App(): JSX.Element {
   const deleteGame = useStore(useCallback((state) => state.deleteGame, []));
   const [confirm, setConfirm] = useState(false);
   const [visible, setVisible] = useState(false);
+
+  useEffect(() => {
+    if (
+      newVersion !== process.env.REACT_APP_VERSION &&
+      window?.matchMedia('(prefers-color-scheme:dark)')?.matches
+    ) {
+      useStore.setState((state) => {
+        state.darkMode = true;
+      });
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
   useEffect(() => {
     if (darkMode) {
@@ -209,7 +221,6 @@ function App(): JSX.Element {
           </main>
         </Sidebar.Pusher>
       </Sidebar.Pushable>
-      <Footer />
       <Effectiveness />
       <ToastContainer
         limit={3}