+++ src/common/EncounterSelector/EncounterSelector.tsx
@@ -28,7 +28,7 @@ function EncounterSelector({
     const foundPokemon = POKEMAP.get(p.pokemon);
     return (
       foundPokemon?.text.toUpperCase().includes(values.search) &&
-      (!!limitGen ? foundPokemon.generation <= limitGen : true) &&
+      (limitGen ? foundPokemon.generation <= limitGen : true) &&
       (values.gens.length > 0 ? values.gens.includes(foundPokemon.generation) : true) &&
       (values.types.length > 0
         ? values.types.includes(foundPokemon.type) || values.types.includes(foundPokemon?.dualtype)