+++ src/common/PokemonSelector/PokemonSelector.test.tsx
@@ -1,16 +1,16 @@
-import { render, fireEvent, screen } from '@testing-library/react';
+import { fireEvent, render, screen } from '@testing-library/react';
 import PokemonSelector from './PokemonSelector';
 import '@testing-library/jest-dom';
 
 describe('Pokemon selector', () => {
   test('Loads Pokemon selector', () => {
     const handlePokemon = jest.fn();
-    const utils = render(
+    render(
       <PokemonSelector handlePokemon={handlePokemon} filter={false}>
         <div>Test</div>
       </PokemonSelector>
     );
-    fireEvent.click(utils.getByText('Test'));
+    fireEvent.click(screen.getByText('Test'));
     const pokemon = screen.getByTestId('poke-Bulbasaur');
     expect(pokemon).toBeVisible();
     fireEvent.click(pokemon);