+++ src/components/index.ts
@@ -7,7 +7,6 @@ export { default as Changelog } from './Changelog/Changelog';
 export { default as Effectiveness } from './Effectiveness/Effectiveness';
 export { default as Export } from './Export/Export';
 export { default as Filter } from './Filter/Filter';
-export { default as Footer } from './Footer/Footer';
 export { default as Import } from './Import/Import';
 export { default as Member } from './Member/Member';
 export { default as Move } from './Move/Move';